# app/docs/swagger.yaml
definitions:
...
  vo.CreateIssueViewReq:
    properties:
      config:
        description: 视图配置
        type: string
      isPrivate:
        description: 是否私有
        type: boolean
      projectId:
        description: 项目 id
        type: integer
      projectObjectTypeId:
        description: 所属任务类型 id：需求、任务、缺陷的 id 值
        type: integer
      remark:
        description: 视图备注
        type: string
      sort:
        description: 视图排序
        type: integer
      type:
        description: 类型，1：表格视图，2：看板视图，3：照片视图
        type: integer
      viewName:
        description: 视图名称
        type: string
    type: object
...
info:
  contact: {}
  description: 极星接口文档.
  title: Polaris Apis
  version: v1.0.0
paths:
...
  /api/rest/issueView/create:
    post:
      consumes:
      - application/json
      description: 新增任务视图
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/vo.CreateIssueViewReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: ""
      security:
      - PM-TOEKN: []
      summary: 新增任务视图
      tags:
      - 任务视图
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"